main: fix vertical-rate units and sign
Two defects in the vertical-rate path:

- ADSBData::getVertVel() returned uint16_t although the field is int16_t, so
  a descent (negative ft/min) was returned as a large positive value. Return
  int16_t.
- The report converted with ft_to_cm (30.48), treating the SBS-1 vertical
  rate as feet rather than feet/minute, so the reported cm/s figure was 60x
  too large. Convert with 30.48 / 60.

The compliance-commit band-aid casts around getVertVel are removed now that
the types line up, and the parser test asserts the signed value directly
instead of casting it back to int16_t.

+++ src/test/test_dump1090.cpp
@@ -94,7 +94,7 @@ TEST_CASE_METHOD(ParserFixture, "MSG type 4 (airborne vel) sets speed, heading a
     CHECK(g_captured->validHeading());
     CHECK(g_captured->getHeading() == 270);
     CHECK(g_captured->validVertVel());
-    CHECK(static_cast<int16_t>(g_captured->getVertVel()) == -1024);
+    CHECK(g_captured->getVertVel() == -1024);
 }
 
 TEST_CASE_METHOD(ParserFixture, "MSG type 6 (surveillance id) sets squawk", "[parser][valid]")
